Add vizzuality endpoints and timeline widget. Misc else.
Big update because it ran away with me while I was working on all this.

Built for the heroku deployment, a big update to the calculation methods
and all of the vizzuality endpoints have been added, plus their schemas.

Not everything works.

Also an env_template file giving a template for your .env file.