Don't use celery singleton with chord callbacks
It looks like this was causing some calculations to fail when the
previous calculation was still in memory?? or something?? and then
the lock prevented new calculations for some time.

The callbacks are cheap enough and unique enough that we probably
don't need to enforce uniqueness on the jobs anyway.