Change bbox from wkt to list of floats. Refactoring.
It's easier for Vizzuality to work with a list of bounding coordinates than well known text
so we'll go back to that.

And we'll take this chance to move things around to make the geocoding dependencies and
standardisation make a liiittle more sense. Taking advantage of util.py and class methods.

It still needs a think and a refactor to be generally useful.