Units conversions
Big refactor of how units work. Add convert_units methods to most Schema.

Now when data is loaded from the database the units can be converted on the fly.

This is done via the wrangle_units decorator around most endpoints. It means that
almost operations inside the API can assume that units are the canonical ones from
 the CLIMADA data API.